Don't crash and burn if message data encoding fails
Apparently we assume that message data is always UTF-8, or it can always
be decoded as ASCII. However, multi-part messages can have parts with
different encodings making lpf panic. The right way (tm) to fix this
would probably be to deal with each part encoding separately, but for
this quick fix we'll just keep the original encoding if re-encoding to
UTF-8 fails.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>