Add option to sort images based on email sender
Instead of just dumping all images in a single folder, create subfolders
for senders to neatly organize the images we received. Multiple copies
of the same image won't be downloaded becase we have a file hash control,
but we'll deal with that later. (Maybe create hardlinks for duplicated
files in different folders?)

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>